python-scripts/requirements.txt
pybit
python-dotenv
datetime